chore(keychain-sdk): move TxWriter to internal pkg
